+++ opencga-server/src/main/java/org/opencb/opencga/server/ws/JobWSServer.java
@@ -70,8 +70,8 @@ public class JobWSServer extends OpenCGAWSServer {
         public InputJob() {
         }
 
-        public InputJob(String name, String toolName, String description, long startTime, long endTime, String commandLine, Status status, int outDirId,
-                        List<Integer> input, Map<String, Object> attributes, Map<String, Object> resourceManagerAttributes) {
+        public InputJob(String name, String toolName, String description, long startTime, long endTime, String commandLine, Status status, long outDirId,
+                        List<Long> input, Map<String, Object> attributes, Map<String, Object> resourceManagerAttributes) {
             this.name = name;
             this.toolName = toolName;
             this.description = description;
@@ -99,9 +99,9 @@ public class JobWSServer extends OpenCGAWSServer {
         public String commandLine;
         public Status status = Status.READY;
         @ApiModelProperty(required = true)
-        public int outDirId;
-        public List<Integer> input;
-        public List<Integer> output;
+        public long outDirId;
+        public List<Long> input;
+        public List<Long> output;
         public Map<String, Object> attributes;
         public Map<String, Object> resourceManagerAttributes;
 
@@ -114,7 +114,7 @@ public class JobWSServer extends OpenCGAWSServer {
     public Response createJobPOST(@ApiParam(value = "studyId", required = true) @QueryParam("studyId") String studyIdStr,
                                   @ApiParam(value = "studies", required = true) InputJob job) {
         try {
-            int studyId = catalogManager.getStudyId(studyIdStr);
+            long studyId = catalogManager.getStudyId(studyIdStr);
             QueryResult<Job> result = catalogManager.createJob(studyId, job.name, job.toolName, job.description,
                     job.execution, job.params, job.commandLine, null, job.outDirId, job.input, job.output, job.attributes,
                     job.resourceManagerAttributes, Job.JobStatus.valueOf(job.status.toString()), job.startTime, job.endTime, queryOptions, sessionId);
@@ -138,10 +138,10 @@ public class JobWSServer extends OpenCGAWSServer {
     ) {
         QueryResult<Job> jobResult;
         try {
-            int studyId = catalogManager.getStudyId(studyIdStr);
+            long studyId = catalogManager.getStudyId(studyIdStr);
             AnalysisJobExecutor analysisJobExecutor;
             String toolName;
-            int toolId = catalogManager.getToolId(toolIdStr);
+            long toolId = catalogManager.getToolId(toolIdStr);
             if (toolId < 0) {
                 analysisJobExecutor = new AnalysisJobExecutor(toolIdStr, execution);    //LEGACY MODE, AVOID USING
                 toolName = toolIdStr;
@@ -151,7 +151,7 @@ public class JobWSServer extends OpenCGAWSServer {
                 toolName = tool.getName();
             }
 
-            List<Integer> inputFiles = new LinkedList<>();
+            List<Long> inputFiles = new LinkedList<>();
             Map<String, List<String>> localParams = new HashMap<>(params);
 
             Execution ex = analysisJobExecutor.getExecution();
@@ -190,7 +190,7 @@ public class JobWSServer extends OpenCGAWSServer {
                 return createErrorResponse("", "Missing output param '" + outputParam + "'");
             }
 
-            int outDirId;
+            long outDirId;
 //            System.out.println("outputParam = " + outputParam);
             if(params.get(outputParam).get(0).equalsIgnoreCase("analysis")){
                 QueryOptions query = new QueryOptions();
@@ -267,7 +267,7 @@ public class JobWSServer extends OpenCGAWSServer {
     @GET
     @Path("/{jobId}/info")
     @ApiOperation(value = "Get job information", position = 2)
-    public Response info(@ApiParam(value = "jobId", required = true) @PathParam("jobId") int jobId) {
+    public Response info(@ApiParam(value = "jobId", required = true) @PathParam("jobId") long jobId) {
         try {
             return createOkResponse(catalogManager.getJob(jobId, queryOptions, sessionId));
         } catch (CatalogException e) {
@@ -278,7 +278,7 @@ public class JobWSServer extends OpenCGAWSServer {
     @GET
     @Path("/{jobId}/visit")
     @ApiOperation(value = "Increment job visits", position = 3)
-    public Response visit(@ApiParam(value = "jobId", required = true) @PathParam("jobId") int jobId) {
+    public Response visit(@ApiParam(value = "jobId", required = true) @PathParam("jobId") long jobId) {
         try {
             return createOkResponse(catalogManager.incJobVisites(jobId, sessionId));
         } catch (CatalogException e) {
@@ -289,13 +289,13 @@ public class JobWSServer extends OpenCGAWSServer {
     @GET
     @Path("/{jobId}/delete")
     @ApiOperation(value = "Delete job", position = 4)
-    public Response delete(@ApiParam(value = "jobId", required = true) @PathParam("jobId") int jobId,
+    public Response delete(@ApiParam(value = "jobId", required = true) @PathParam("jobId") long jobId,
                            @ApiParam(value = "deleteFiles", required = true) @DefaultValue("true") @QueryParam("deleteFiles") boolean deleteFiles) {
         try {
             List<QueryResult> results = new LinkedList<>();
             if (deleteFiles) {
                 QueryResult<Job> jobQueryResult = catalogManager.getJob(jobId, null, sessionId);
-                for (Integer fileId : jobQueryResult.getResult().get(0).getOutput()) {
+                for (Long fileId : jobQueryResult.getResult().get(0).getOutput()) {
                     QueryResult queryResult = catalogManager.deleteFile(fileId, sessionId);
                     results.add(queryResult);
                 }
